Remove TimeZone requirement and fix redis client

+++ apps/farmbot/lib/mix/tasks/upload.ex
@@ -4,12 +4,20 @@ defmodule Mix.Tasks.Farmbot.Upload do
   alias Mix.Tasks.Firmware.Push
   @shortdoc "Uploads a file to a url"
   def run(args) do
-    img_path = Mix.Project.config[:images_path]
-    fw_file = Path.join(img_path, "farmbot.fw")
+    otp_app = Mix.Project.config[:app]
+    target = Mix.Project.config[:target]
+    # TODO(Connor) change this back when nerves is stable again
+    # img_path = Mix.Project.config[:images_path]
+    # fw_file = Path.join(img_path, "#{otp_app}.fw")
+    fw_file = Path.join(["_images", "#{target}", "#{otp_app}.fw"])
     unless File.exists?(fw_file) do
        raise "Could not find Firmware!"
     end
     ip_address = List.first(args)
-    Push.run([ip_address, "--firmware", "#{fw_file}", "--reboot", "true"])
+    if Code.ensure_loaded?(Push) do
+      Push.run([ip_address, "--firmware", "#{fw_file}", "--reboot", "true"])
+    else
+      Mix.raise("ERRRRR?")
+    end
   end
 end
